requirement.txt
flask
requests
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
langchain-core
langsmith
PyPDF2
